+++ apps/site/src/app/ecosystem-explorer/page.tsx
@@ -1,62 +1,30 @@
-import dynamic from 'next/dynamic'
-
-import { BookOpen } from '@phosphor-icons/react/dist/ssr'
-
 import type { AsyncNextServerSearchParams } from '@/types/searchParams'
 
 import { PATHS } from '@/constants/paths'
-import { CATEGORY_KEY } from '@/constants/searchParams'
 
 import { graphicsData } from '@/data/graphicsData'
 
-import { buildImageSizeProp } from '@/utils/buildImageSizeProp'
 import { createMetadata } from '@/utils/createMetadata'
-import { entryMatchesCategoryQuery } from '@/utils/filterUtils'
-import { findOrThrow } from '@/utils/findOrThrow'
 import { getFrontmatter } from '@/utils/getFrontmatter'
 import { getSortOptions } from '@/utils/getSortOptions'
 
 import { BaseFrontmatterSchema } from '@/schemas/FrontmatterSchema'
 
-import { useFilter } from '@/hooks/useFilter'
-import { usePagination } from '@/hooks/usePagination'
-import { useSearch } from '@/hooks/useSearch'
-import { useSort } from '@/hooks/useSort'
-
-import { Card } from '@/components/Card'
-import { CardGrid } from '@/components/CardGrid'
 import { CTASection } from '@/components/CTASection'
-import { FilterContainer } from '@/components/FilterContainer'
-import { NoSearchResultsMessage } from '@/components/NoSearchResultsMessage'
 import { PageHeader } from '@/components/PageHeader'
 import { PageLayout } from '@/components/PageLayout'
 import { PageSection } from '@/components/PageSection'
-import { ResultsAndReset } from '@/components/ResultsAndReset'
-import { Search } from '@/components/Search'
-import { Sort } from '@/components/Sort'
 import { StructuredDataScript } from '@/components/StructuredDataScript'
 
-import { CategoryFilters } from './components/CategoryFilters'
-import { CategoryFiltersSlider } from './components/CategoryFiltersSlider'
+import { EcosystemExplorerContent } from './components/EcosystemExplorerContent'
 import { ecosystemProjectsSortConfigs } from './constants/sortConfigs'
-import { useEcosystemCategoryTree } from './hooks/useEcosystemCategoryTree'
 import { generateStructuredData } from './utils/generateStructuredData'
 import { getEcosystemCMSCategories } from './utils/getEcosystemCMSCategories'
 import { getEcosystemProjectsData } from './utils/getEcosystemProjectData'
-import { parseCategoryQueryParam } from './utils/parseCategoryQueryParam'
-
-const NoSSRPagination = dynamic(
-  () => import('@/components/Pagination').then((module) => module.Pagination),
-  { ssr: false },
-)
 
 type Props = {
   searchParams: AsyncNextServerSearchParams
 }
-
-const ecosystemProjects = getEcosystemProjectsData()
-
-const sortOptions = getSortOptions(ecosystemProjectsSortConfigs)
 
 const { header, seo } = getFrontmatter({
   path: PATHS.ECOSYSTEM_EXPLORER,
@@ -73,42 +41,16 @@ export const metadata = createMetadata({
 })
 
 const categories = getEcosystemCMSCategories()
+const ecosystemProjects = getEcosystemProjectsData()
+const sortOptions = getSortOptions(ecosystemProjectsSortConfigs)
 
 export default async function EcosystemExplorer(props: Props) {
   const searchParams = await props.searchParams
-
-  const { searchQuery, searchResults } = useSearch({
-    searchParams,
-    entries: ecosystemProjects,
-    searchBy: ['title', 'description'],
-  })
-
-  const { sortQuery, sortedResults, defaultSortQuery } = useSort({
-    searchParams,
-    entries: searchResults,
-    configs: ecosystemProjectsSortConfigs,
-    defaultsTo: 'a-z',
-  })
-
-  const { filteredEntries } = useFilter({
-    entries: sortedResults,
-    filterQuery: parseCategoryQueryParam(searchParams, CATEGORY_KEY),
-    filterFn: entryMatchesCategoryQuery,
-  })
-
-  const categoryTree = useEcosystemCategoryTree({
-    entries: sortedResults,
-    categories,
-  })
-
-  const { currentPage, pageCount, paginatedResults } = usePagination({
-    searchParams,
-    entries: filteredEntries,
-  })
+  const structuredData = generateStructuredData(seo)
 
   return (
     <PageLayout>
-      <StructuredDataScript structuredData={generateStructuredData(seo)} />
+      <StructuredDataScript structuredData={structuredData} />
       <PageHeader
         title={header.title}
         description={header.description}
@@ -124,100 +66,12 @@ export default async function EcosystemExplorer(props: Props) {
         title="Ecosystem Projects"
         description="Discover the diverse landscape of projects in the Filecoin ecosystem. Inclusion in the Filecoin Ecosystem Explorer is not an endorsement of any project, any company, or any company’s product or services."
       >
-        <FilterContainer>
-          <FilterContainer.ResultsAndCategory
-            gapSize="wide"
-            results={<ResultsAndReset results={filteredEntries.length} />}
-            category={<CategoryFilters categories={categoryTree} />}
-          />
-          <FilterContainer.MainWrapper>
-            <FilterContainer.DesktopFilters
-              searchComponent={<Search query={searchQuery} />}
-              sortComponent={
-                <Sort
-                  query={sortQuery}
-                  options={sortOptions}
-                  defaultQuery={defaultSortQuery}
-                />
-              }
-            />
-
-            <FilterContainer.MobileFiltersAndResults
-              searchComponent={<Search query={searchQuery} />}
-              filterComponents={[
-                <CategoryFiltersSlider
-                  key="category"
-                  categories={categoryTree}
-                />,
-              ]}
-              sortComponent={
-                <Sort
-                  query={sortQuery}
-                  options={sortOptions}
-                  defaultQuery={defaultSortQuery}
-                />
-              }
-            />
-            <FilterContainer.ContentWrapper>
-              {filteredEntries.length === 0 ? (
-                <NoSearchResultsMessage />
-              ) : (
-                <>
-                  <CardGrid cols="smTwo">
-                    {paginatedResults.map((project, i) => {
-                      const {
-                        slug,
-                        title,
-                        description,
-                        image,
-                        category: categoryId,
-                      } = project
-
-                      const isFirstTwoImages = i < 2
-                      const category = findOrThrow(
-                        categories,
-                        ({ value }) => value === categoryId,
-                      )
-
-                      return (
-                        <Card
-                          key={slug}
-                          title={title}
-                          description={description}
-                          tags={[{ text: category.label }]}
-                          cta={{
-                            href: `${PATHS.ECOSYSTEM_EXPLORER.path}/${slug}`,
-                            text: 'Learn More',
-                            icon: BookOpen,
-                          }}
-                          image={{
-                            ...(image || graphicsData.imageFallback.data),
-                            alt: '',
-                            objectFit: 'contain',
-                            padding: Boolean(image),
-                            priority: isFirstTwoImages,
-                            sizes: buildImageSizeProp({
-                              startSize: '100vw',
-                              sm: '320px',
-                              md: '440px',
-                              lg: '280px',
-                            }),
-                          }}
-                        />
-                      )
-                    })}
-                  </CardGrid>
-                  <FilterContainer.PaginationWrapper>
-                    <NoSSRPagination
-                      pageCount={pageCount}
-                      currentPage={currentPage}
-                    />
-                  </FilterContainer.PaginationWrapper>
-                </>
-              )}
-            </FilterContainer.ContentWrapper>
-          </FilterContainer.MainWrapper>
-        </FilterContainer>
+        <EcosystemExplorerContent
+          searchParams={searchParams}
+          ecosystemProjects={ecosystemProjects}
+          categories={categories}
+          sortOptions={sortOptions}
+        />
       </PageSection>
 
       <CTASection
